+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### ensmallen ?.??.?
 ###### ????-??-??
+  * Allow callback instantiation for SGD based optimizer 
+    ([#138](https://github.com/mlpack/ensmallen/pull/155)).
+
   * Minor test stability fixes on i386
     ([#156](https://github.com/mlpack/ensmallen/pull/156)).